refactor: move cli to pkg directory
