+++ pkg/cli/license.go
